refactor(crane): hoist the crane CLI wrapper into internal/crane
