+++ cmd/c8s/install_test.go
@@ -1351,30 +1351,6 @@ func TestBuildDigestArgsLeavesOtherResolveErrorsUnhinted(t *testing.T) {
 	}
 }
 
-// isImageNotFound keys the tag-coupling guidance to the registry's own
-// missing-reference error codes, so auth and network failures never
-// masquerade as a missing tag.
-func TestIsImageNotFound(t *testing.T) {
-	tests := []struct {
-		name string
-		err  error
-		want bool
-	}{
-		{name: "missing tag", err: errors.New("crane digest: MANIFEST_UNKNOWN: manifest unknown"), want: true},
-		{name: "missing repository", err: errors.New("crane digest: NAME_UNKNOWN: repository name not known to registry"), want: true},
-		{name: "auth failure", err: errors.New("crane digest: UNAUTHORIZED: authentication required"), want: false},
-		{name: "network failure", err: errors.New("dial tcp: lookup ghcr.io: no such host"), want: false},
-		{name: "nil", err: nil, want: false},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := isImageNotFound(tt.err); got != tt.want {
-				t.Fatalf("isImageNotFound(%v) = %t, want %t", tt.err, got, tt.want)
-			}
-		})
-	}
-}
-
 // chartComponents reads the component set from the chart's values.yaml; this
 // asserts the parse against the embedded chart so the install-time list cannot
 // silently diverge from what the chart declares.
